Web page assignment
It took me the better part of the day to fool around with this code.  I had to watch the videos multiple times to see if i had inputed any errors and found only minor input differences.  I had to mess with the instructors code a bit for the pictures to come out the way it looks.